refactor: separate logic and UI, add console logs in RestClient
